+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpX509HashTest.kt
@@ -22,6 +22,7 @@ import io.kotest.matchers.collections.shouldBeSingleton
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.types.shouldBeInstanceOf
 
+@Suppress("unused")
 val OpenId4VpX509HashTest by testSuite {
 
     withFixtureGenerator(suspend {
@@ -94,7 +95,7 @@ val OpenId4VpX509HashTest by testSuite {
             it.verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode()).getOrThrow()
                 .vpTokenValidationResult.shouldNotBeNull().getOrThrow()
                 .shouldBeInstanceOf<VpTokenValidationResultDCQL>().apply {
-                    allValidationResults.values
+                    credentialQueryResponseValidations.values
                         .shouldBeSingleton().first().shouldBeSingleton().first().getOrThrow()
                         .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
                         .reconstructedJsonObject[CLAIM_GIVEN_NAME].shouldNotBeNull()
@@ -133,10 +134,11 @@ val OpenId4VpX509HashTest by testSuite {
             it.verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode()).getOrThrow()
                 .vpTokenValidationResult.shouldNotBeNull().getOrThrow()
                 .shouldBeInstanceOf<VpTokenValidationResultDCQL>().apply {
-                    allValidationResults.values
+                    credentialQueryResponseValidations.values
                         .shouldBeSingleton().first()
-                        .shouldBeSingleton().first().getOrThrow().shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
-                        .reconstructed[CLAIM_GIVEN_NAME].shouldNotBeNull()
+                        .shouldBeSingleton().first().getOrThrow()
+                        .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
+                        .reconstructedJsonObject[CLAIM_GIVEN_NAME].shouldNotBeNull()
                 }
 
         }